pyproject.toml
[project]
name = "juniorweb-downloader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "lxml-html-clean>=0.4.3",
    "python-dotenv>=1.2.1",
    "python-magic>=0.4.27",
    "requests>=2.32.5",
    "requests-html>=0.10.0",
    "urllib3>=2.6.3",
]
